Change the files
Delete AVL_test header and cpp
