+++ Projects-2016-2017/AVL Ionescu Iulian/AVL.h
@@ -31,5 +31,8 @@ void search_node(NodeAVL *node, int value,int *found);
 NodeAVL* findmin(NodeAVL* node);
 NodeAVL* removemin(NodeAVL* node);
 NodeAVL* remove(NodeAVL* node, int k);
-
+int test_insertAVLNode(char *file,int arborescenta);
+int test_deleteAVLNode(char *file,int arborescenta);
+int isAVLTree(NodeAVL *root);
+void buildInput (int size, int min, int max, char file[]);
 #endif
