--- Projects-2016-2017/AVL Ionescu Iulian/AVL_Test.h
@@ -1,5 +1,0 @@
-#include "AVL.h"
-int test_insertAVLNode(char *file,int arborescenta);
-int test_deleteAVLNode(char *file,int arborescenta);
-int isAVLTree(NodeAVL *root);
-void buildInput (int size, int min, int max, char file[]);
