--- Projects-2016-2017/AVL Ionescu Iulian/avl/date.txt
@@ -1,20 +1,0 @@
-3
-19
-12
-9
-6
-4
-17
-13
-14
-10
-1
-5
-15
-11
-2
-16
-18
-8
-7
-20
