+++ Projects-2016-2017/AVL Ionescu Iulian/date.txt
@@ -1,10 +1,10 @@
-1
-25
-10
 16
-22
-6
-17
-7
-23
+26
+25
+8
 2
+23
+7
+6
+5
+18
