+++ Projects-2016-2017/AVL Ionescu Iulian/input.txt
@@ -1,8 +1,8 @@
-I 7
-I 2
-I 8
-I 1
-I 4
-I 3
-I 5
+7
+2
+8
+1
+4
+3
+5
